third_party/blink/web_tests/fast/css/invalidation/removed-hover-shadow-rule-expected.txt

CONSOLE WARNING: line 34: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Removing a shadow tree stylesheet should cause an invalidation set update.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.internals is defined.
PASS window.eventSender is defined.
PASS getComputedStyle(match).backgroundColor is red
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(match).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

Hover this text ... and this should have a green background.
